Replace Dagger with Koin
