+++ app/build.gradle
@@ -54,10 +54,6 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-gson:$retrofit_version"
     implementation 'com.squareup.okhttp3:logging-interceptor:3.10.0'
 
-    // Dagger
-    def dagger_version = "2.19"
-    api "com.google.dagger:dagger:$dagger_version"
-    kapt "com.google.dagger:dagger-compiler:$dagger_version"
-    compileOnly "com.squareup.inject:assisted-inject-annotations-dagger2:0.3.3"
-    kapt "com.squareup.inject:assisted-inject-processor-dagger2:0.3.3"
+    // Koin
+    implementation 'org.koin:koin-android-viewmodel:2.0.1'
 }
