+++ app/src/main/java/co/rahulchowdhury/elly/ElephantApp.kt
@@ -1,18 +1,24 @@
 package co.rahulchowdhury.elly
 
 import android.app.Application
-import co.rahulchowdhury.elly.injection.component.AppComponent
-import co.rahulchowdhury.elly.injection.component.DaggerAppComponent
-import co.rahulchowdhury.elly.injection.module.AppModule
+import co.rahulchowdhury.elly.di.module.elephantModule
+import co.rahulchowdhury.elly.di.module.networkModule
+import org.koin.android.ext.koin.androidContext
+import org.koin.core.context.startKoin
 
 class ElephantApp : Application() {
-    lateinit var appComponent: AppComponent
+    //lateinit var appComponent: AppComponent
 
     override fun onCreate() {
         super.onCreate()
-        appComponent = DaggerAppComponent
-            .builder()
-            .appModule(AppModule(this))
-            .build()
+//        appComponent = DaggerAppComponent
+//            .builder()
+//            .appModule(AppModule(this))
+//            .build()
+
+        startKoin {
+            androidContext(this@ElephantApp)
+            modules(listOf(networkModule, elephantModule))
+        }
     }
 }
