+++ app/src/main/java/co/rahulchowdhury/elly/di/module/NetworkModule.kt
@@ -1,0 +1,45 @@
+package co.rahulchowdhury.elly.di.module
+
+import co.rahulchowdhury.elly.data.source.remote.elephant.ElephantApiService
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import org.koin.dsl.module
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+val networkModule = module {
+    single { provideHttpLoggingInterceptor() }
+    single { provideOkHttpClient(get()) }
+    single { provideRetrofit(get()) }
+    single { provideElephantApiService(get()) }
+}
+
+fun provideHttpLoggingInterceptor(): HttpLoggingInterceptor {
+    val httpLoggingInterceptor = HttpLoggingInterceptor()
+    httpLoggingInterceptor.level = HttpLoggingInterceptor.Level.BODY
+
+    return httpLoggingInterceptor
+}
+
+fun provideOkHttpClient(
+    httpLoggingInterceptor: HttpLoggingInterceptor
+): OkHttpClient {
+    val okHttpClientBuilder = OkHttpClient.Builder()
+
+    okHttpClientBuilder.addInterceptor(httpLoggingInterceptor)
+    return okHttpClientBuilder.build()
+}
+
+fun provideRetrofit(
+    okHttpClient: OkHttpClient
+): Retrofit =
+    Retrofit.Builder()
+        .client(okHttpClient)
+        .addConverterFactory(GsonConverterFactory.create())
+        .baseUrl("https://elephant-api.herokuapp.com")
+        .build()
+
+fun provideElephantApiService(
+    retrofit: Retrofit
+): ElephantApiService =
+    retrofit.create(ElephantApiService::class.java)
