--- app/src/main/java/co/rahulchowdhury/elly/injection/base/ViewModelAssistedFactory.kt
@@ -1,8 +1,0 @@
-package co.rahulchowdhury.elly.injection.base
-
-import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
-
-interface ViewModelAssistedFactory<T : ViewModel> {
-    fun create(savedStateHandle: SavedStateHandle): T
-}
