--- app/src/main/java/co/rahulchowdhury/elly/injection/base/ViewModelKey.kt
@@ -1,11 +1,0 @@
-package co.rahulchowdhury.elly.injection.base
-
-import androidx.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-@MustBeDocumented
-@Target(AnnotationTarget.FUNCTION)
-@Retention(AnnotationRetention.RUNTIME)
-@MapKey
-annotation class ViewModelKey(val value: KClass<out ViewModel>)
