--- app/src/main/java/co/rahulchowdhury/elly/injection/factory/ViewModelFactory.kt
@@ -1,26 +1,0 @@
-package co.rahulchowdhury.elly.injection.factory
-
-import android.os.Bundle
-import androidx.lifecycle.AbstractSavedStateViewModelFactory
-import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
-import androidx.savedstate.SavedStateRegistryOwner
-import co.rahulchowdhury.elly.injection.base.ViewModelAssistedFactory
-import javax.inject.Inject
-
-class ViewModelFactory @Inject constructor(
-    private val viewModelMap: MutableMap<Class<out ViewModel>, ViewModelAssistedFactory<out ViewModel>>,
-    owner: SavedStateRegistryOwner,
-    defaultArgs: Bundle?
-) : AbstractSavedStateViewModelFactory(owner, defaultArgs) {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel?> create(
-        key: String,
-        modelClass: Class<T>,
-        handle: SavedStateHandle
-    ): T {
-        return viewModelMap[modelClass]?.create(handle) as? T
-            ?: throw IllegalStateException("Unknown ViewModel class")
-    }
-}
