--- app/src/main/java/co/rahulchowdhury/elly/injection/module/NetworkModule.kt
@@ -1,51 +1,0 @@
-package co.rahulchowdhury.elly.injection.module
-
-import co.rahulchowdhury.elly.data.source.remote.elephant.ElephantApiService
-import dagger.Module
-import dagger.Provides
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-class NetworkModule {
-    @Provides
-    @Singleton
-    fun provideHttpLoggingInterceptor(): HttpLoggingInterceptor {
-        val httpLoggingInterceptor = HttpLoggingInterceptor()
-        httpLoggingInterceptor.level = HttpLoggingInterceptor.Level.BODY
-
-        return httpLoggingInterceptor
-    }
-
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(
-        httpLoggingInterceptor: HttpLoggingInterceptor
-    ): OkHttpClient {
-        val okHttpClientBuilder = OkHttpClient.Builder()
-
-        okHttpClientBuilder.addInterceptor(httpLoggingInterceptor)
-        return okHttpClientBuilder.build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideRetrofit(
-        okHttpClient: OkHttpClient
-    ): Retrofit =
-        Retrofit.Builder()
-            .client(okHttpClient)
-            .addConverterFactory(GsonConverterFactory.create())
-            .baseUrl("https://elephant-api.herokuapp.com")
-            .build()
-
-    @Provides
-    @Singleton
-    fun provideElephantApiService(
-        retrofit: Retrofit
-    ): ElephantApiService =
-        retrofit.create(ElephantApiService::class.java)
-}
