--- app/src/main/java/co/rahulchowdhury/elly/injection/module/ViewModelAssistedFactoriesModule.kt
@@ -1,8 +1,0 @@
-package co.rahulchowdhury.elly.injection.module
-
-import com.squareup.inject.assisted.dagger2.AssistedModule
-import dagger.Module
-
-@AssistedModule
-@Module(includes = [AssistedInject_ViewModelAssistedFactoriesModule::class])
-abstract class ViewModelAssistedFactoriesModule
