--- app/src/main/java/co/rahulchowdhury/elly/injection/scope/App.kt
@@ -1,7 +1,0 @@
-package co.rahulchowdhury.elly.injection.scope
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.SOURCE)
-annotation class App
