+++ app/src/main/java/co/rahulchowdhury/elly/ui/profile/ElephantProfileViewModel.kt
@@ -4,9 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import co.rahulchowdhury.elly.data.model.local.Elephant
 import co.rahulchowdhury.elly.data.repo.ElephantRepository
-import javax.inject.Inject
 
-class ElephantProfileViewModel @Inject constructor(
+class ElephantProfileViewModel(
     elephantRepository: ElephantRepository
 ) : ViewModel() {
     private val elephantName: String = "Arjuna"
