+++ app/src/main/res/layout/elephant_profile_fragment.xml
@@ -7,6 +7,7 @@
         tools:context=".ui.profile.ElephantProfileFragment">
 
     <TextView
+            android:id="@+id/apiResponse"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:text="Hello"/>
